Some changes made.
Now style is a CanvasObject property. it is now inherited for every child prototype.

CanvasObject().prepare() is now used to apply pre-rendering to the CanvasObject().update(), instead off being written for every single new object created.

CanvasObject().updateChild() is now called for children update in the end of CanvasObject().update().

To prevent bugs, CanvasObject.rotOffX() will now check if orbit is 0 from now on to calculate the object position. Any tips to improve are welcome.

CanvasObject and Canvas setters for width and height were modified to be more readable and to give developers outputs of values warnings.

Every child of CanvasObject was modified to match the new CanvasObject().style, CanvasObject.prepare() and CanvasObject.updateChild() property.

Added new comments. Any comment improvement is welcome.